Switching source format to '3.0 (quilt)' and changing debian/rules to use the 'dh $@' generic rule plus overrides where needed.
Now that the upstream handling of gettext files has been fixed (cf.
the previous patch) this switching is feasible and results in a short
and elegant debian/rules file.

Those that use dpkg-buildpackage to build must add the -tc option in
order for final 'make clean' to get performed with upstream makefiles
patched. This will not be needed anymore when upstream incorporates
the previous patch.