Depend on perl rather than merely suggesting it
Contrary to what I said in commit d49dff5b12280, perl is not only
required by backup-manager-upload, but also by backup-manager-purge as
part of the normal backup process, and not having it installed means
that old archives will not be removed.

Don't talk about perl in README.Debian, since it is no longer
optionally.